requirements.txt
Flask==3.0.3, 
Flask-SQLAlchemy==3.1.1, 
Flask-Migrate==4.0.7, 
Flask-CORS==4.0.1, 
SQLAlchemy==2.0.31, 
python-dotenv==1.0.1, 
pandas==2.2.2, 
numpy==1.26.4, 
scikit-learn==1.5.1, 
nltk==3.8.1, 
pytest==8.2.2, 
pytest-flask==1.3.0